extract module: get_object::status
